filter: support parsing numeric index
The intention of this is to eventually allow folks to query and delete
specific indicies within a list type. Additionally, this adds further
restrictions to prevent parsing properties and names that start with
digits which is not allowed in HCL.

This should not be released as a new version yet as it is not used
in any of the subcommands. I mainly opened this PR separate from
the implementation to get feedback before I spend time on the
implementation.